refactor(evm): split sentrix-precompiles crate (Tier 2 #9)
Third crate split overnight. Tier 2 pure-data extraction — precompile
address constants + is_sentrix_precompile() helper.

Moved from crates/sentrix-evm/src/precompiles.rs into new crate
crates/sentrix-precompiles/:
  - STAKING_PRECOMPILE (address 0x0100)
  - SLASHING_PRECOMPILE (address 0x0101)
  - is_sentrix_precompile(addr) -> bool

Rationale: future sentrix-sdk + governance tooling want to reference
these canonical addresses without pulling the full sentrix-evm stack
(revm 37 + 10+ transitive alloy crates).

Consensus note: the numeric addresses are part of Sentrix's contract
surface post-Voyager-EVM. Smart contracts that encode these addresses
as constants depend on them being stable. Pure `git mv` preserves byte
values exactly — zero consensus impact.

Back-compat shim at crates/sentrix-evm/src/precompiles.rs re-exports
all three items so existing `sentrix_evm::precompiles::*` imports keep
working unchanged. Follow-up PR will migrate direct call sites + delete
the shim.

sentrix-evm now depends on sentrix-precompiles. 2 unit tests moved
with the code; both pass unchanged.
